Move API into API module
